Implement entity to fetch countries from the API
This entity will take care of fetching the countries, and will do
so by making use of URLSession, which has been made to conform to
our custom protocol.

URLSession is straight forward enough for our use-case as an
HTTP client, no need for Alamofire.

Endpoints have been constructed by using URLComponents, as that
allows us to be explicit about the API and potentially extend it
in the future by means of the `url(relativeTo:)` method, which can
take a baseURL.