Display countries retrieved from restcountries.com
This commit introduces the Main view which displays the countries
returned by https://restcountries.com.

As we interface with a network resource, the network call can fail
for a variety of reasons. The UI will display that state accordingly
and respond to it.

Because data may change, and the current implementation fetches all
countries upon display, I've also opted to add a `refreshable`,
granting the possibility to refresh the data.
To avoid losing the data again, the `onAppear` checks whether any
countries have been fetched already, and fetches them only if needed.

Dependencies are injected through a Composition root, a style which
declares dependencies in one specific location, making sure that
other components are able to fetch it, but don't need the internals
on how to create those dependencies.