Provide detail view to consult Country information
This commit introduces the ability to get some more informatino about
a country. A user can tap the row and will be direct to a detail
view.

The detail view lists the country's name in the navigation bar as
context for the user, as well as the flag in emoji-style.
To make it visual, the PNG variant is also fetched and displayed.

Other fields are visualised through the use of so-called 'metadata'
fields which allow the View to loop over those fields and create the
required `Text` elements where needed.

In doing so I've opted to use a `Router` pattern, to keep the
navigation outside of our View insofar that is possible with the
`NavigationStack`.
This method allows to register a class which is then used to
determine which View to navigate to. The Composition root again
allows us to supply the required dependencies without having to put
that knowledge inside of any other component.